Get a graph

This function will generate a graph to visualize the predicted gene regulatory network

GRNPlot(
  df.grn,
  tfs.use = NULL,
  show.tf.labels = TRUE,
  tfs.timepoint = NULL,
  genes.cluster = NULL,
  genes.use = NULL,
  genes.highlight = NULL,
  cols.highlight = "#984ea3",
  seed = 42,
  plot.importance = TRUE,
  min.importance = 2,
  remove.isolated = FALSE
)

Arguments

df.grn

A data frame representing predicted network

tfs.timepoint

Time points of TFs

genes.cluster

A data frame containing clustering results of genes

genes.highlight

A string list to include gene names for plotting

cols.highlight

Color code for highlighted genes

seed

Random seet

plot.importance

Whether or not plot the scatter plot to visualize importance score of each TF

min.importance

The minimum importance score for showing the TF labels.

Value

A ggplot object