Get list of paired cells

This function takes in the output of the fullmatch function and sort the results into a list of ATAC-RNA pairs. It was modified from https://github.com/buenrostrolab/stimATAC_analyses_code/blob/master/R/optMatching_functions.R
Author: Vinay Kartha, Yan Hu
Contact: <vinay_kartha@g.harvard.edu>
Affiliation: Buenrostro Lab, Department of Stem Cell and Regenerative Biology, Harvard University

GetPairList(cell_matches, ATAC_barcodes, RNA_barcodes)

Arguments

cell_matches

The output object of fullmatch in the package optmatch

ATAC_barcodes

Barcode of ATAC cells, must match the order of IDs in cell_matches.

RNA_barcodes

Barcode of RNA cells, must match the order of IDs in cell_matches.

Value

A data frame containing the cell pairs