Plot trajectory heatmap

This function generates a heatmap to visualize the data along inferred trajectory. It was modified from the package ArchR plotTrajectoryHeatmap. For more details, check here https://www.archrproject.com/reference/plotTrajectoryHeatmap.html.

TrajectoryHeatmap(
  trajectory,
  varCutOff = 0.9,
  maxFeatures = 25000,
  scaleRows = TRUE,
  rowOrder = NULL,
  limits = c(-1.5, 1.5),
  labelRows = FALSE,
  pal = NULL,
  labelMarkers = NULL,
  labelTop = 50,
  name = "Heatmap",
  returnMatrix = FALSE
)

Arguments

trajectory

A SummarizedExperiment object that results from calling GetTrajectory

varCutOff

The "Variance Quantile Cutoff" to be used for identifying the top variable features across the given trajectory. Only features with a variance above the provided quantile will be retained.

maxFeatures

The maximum number of features, ordered by variance, to consider from useMatrix when generating a trajectory

scaleRows

A boolean value that indicates whether row-wise z-scores should be computed on the matrix

rowOrder

If wanting to set the order of rows to be plotted, the indices (integer or character correpsonding to rownmaes) can be provided here.

limits

A numeric vector of two numbers that represent the lower and upper limits of the heatmap color scheme.

labelRows

A boolean value that indicates whether all rows should be labeled on the side of the heatmap.

pal

A custom continuous palette used to override the default continuous palette for the heatmap.

labelMarkers

A character vector listing the rownames that should be labeled on the side of the heatmap.

labelTop

A number indicating how many of the top N features, based on variance of the matrix

name

Name of the matrix

returnMatrix

A boolean value that indicates whether the final heatmap matrix should be returned in lieu of plotting the actual heatmap.

Value

a heatmap